Undo move formatting into one big lump
