+++ mimsa.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 53eb04979eb92a2d34668a2ccec3b149088ce79fd72bf589dbd6ae4311dfb1db
+-- hash: 33f17bd8699615f828000a63f7d1eb73e0af3d84c8c36ecaf328260e3a59e482
 
 name:           mimsa
 version:        0.1.0.0
@@ -40,12 +40,6 @@ library
       Language.Mimsa.Parser.Language
       Language.Mimsa.Parser.Parser
       Language.Mimsa.Printer
-      Language.Mimsa.Printer.Construct
-      Language.Mimsa.Printer.MonoType
-      Language.Mimsa.Printer.Name
-      Language.Mimsa.Printer.TypeError
-      Language.Mimsa.Printer.TypeName
-      Language.Mimsa.Printer.Variable
       Language.Mimsa.Project
       Language.Mimsa.Project.Persistence
       Language.Mimsa.Project.Usages
