--- src/Language/Mimsa/Printer/Construct.hs
@@ -1,7 +1,0 @@
-module Language.Mimsa.Printer.Construct where
-
-import Data.Text.Prettyprint.Doc
-import Language.Mimsa.Types.Construct
-
-renderConstruct :: Construct -> Doc ann
-renderConstruct = pretty . getConstruct
