+++ src/Language/Mimsa/Types/Bindings.hs
@@ -1,10 +1,14 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Language.Mimsa.Types.Bindings where
 
 import qualified Data.Aeson as JSON
 import Data.Map (Map)
+import qualified Data.Map as M
+import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Name
 
@@ -19,3 +23,6 @@ newtype Bindings = Bindings {getBindings :: Map Name ExprHash}
       JSON.FromJSON,
       JSON.ToJSON
     )
+
+instance Printer Bindings where
+  prettyPrint (Bindings b) = "{ " <> T.intercalate ", " (prettyPrint <$> M.keys b) <> " }"
