+++ src/Language/Mimsa/Types/DataType.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Language.Mimsa.Types.DataType
   ( DataType (..),
@@ -9,7 +10,10 @@ where
 
 import qualified Data.Aeson as JSON
 import Data.Map (Map)
+import qualified Data.Map as M
+import Data.Text.Prettyprint.Doc
 import GHC.Generics
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.TypeName
@@ -23,3 +27,24 @@ data DataType
         dtConstructors :: Map Construct [TypeName]
       }
   deriving (Eq, Ord, Show, Generic, JSON.FromJSON, JSON.ToJSON)
+
+instance Printer DataType where
+  prettyDoc = renderDataType
+
+renderDataType :: DataType -> Doc ann
+renderDataType (DataType construct' vars' constructors') =
+  "type " <+> prettyDoc construct'
+    <+> printVars vars'
+    <+> if M.null constructors'
+      then space
+      else
+        sep $
+          zipWith
+            (<+>)
+            ("=" : repeat "|")
+            (printCons <$> M.toList constructors')
+  where
+    printVars [] = mempty
+    printVars as = sep $ renderName <$> as
+    printCons (consName, args) =
+      sep $ [prettyDoc consName] <> (prettyDoc <$> args)
