+++ src/Language/Mimsa/Types/Error.hs
@@ -1,6 +1,9 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Language.Mimsa.Types.Error where
 
 import Data.Text (Text)
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.InterpreterError
 import Language.Mimsa.Types.ResolverError
 import Language.Mimsa.Types.TypeError
@@ -14,3 +17,11 @@ data Error
   | ParseErr Text
   | OtherError Text
   deriving (Eq, Ord, Show)
+
+instance Printer Error where
+  prettyPrint (TypeErr t) = "TypeError: " <> prettyPrint t
+  prettyPrint (ResolverErr a) = "ResolverError: " <> prettyPrint a
+  prettyPrint (InterpreterErr a) = "InterpreterError: " <> prettyPrint a
+  prettyPrint (UsageErr a) = "UsageError: " <> prettyPrint a
+  prettyPrint (ParseErr a) = "ParseError: " <> a
+  prettyPrint (OtherError a) = "OtherError: " <> a
