+++ src/Language/Mimsa/Types/Literal.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Language.Mimsa.Types.Literal
   ( Literal (..),
@@ -8,7 +9,9 @@ module Language.Mimsa.Types.Literal
 where
 
 import qualified Data.Aeson as JSON
+import Data.Text.Prettyprint.Doc
 import GHC.Generics
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.StringType
 
 -------
@@ -26,3 +29,13 @@ data Literal
       JSON.FromJSON,
       JSON.ToJSON
     )
+
+instance Printer Literal where
+  prettyDoc = renderLiteral
+
+renderLiteral :: Literal -> Doc ann
+renderLiteral (MyInt i) = pretty i
+renderLiteral (MyBool True) = "True"
+renderLiteral (MyBool False) = "False"
+renderLiteral (MyString str) = dquotes $ prettyDoc str
+renderLiteral MyUnit = "Unit"
