+++ src/Language/Mimsa/Types/ResolvedDeps.hs
@@ -1,7 +1,11 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Language.Mimsa.Types.ResolvedDeps where
 
 import Data.Map (Map)
 import qualified Data.Map as M
+import Data.Text.Prettyprint.Doc
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.StoreExpression
@@ -15,3 +19,11 @@ newtype ResolvedDeps
 
 hasNoDeps :: ResolvedDeps -> Bool
 hasNoDeps (ResolvedDeps m) = M.size m == 0
+
+instance Printer ResolvedDeps where
+  prettyDoc (ResolvedDeps deps) =
+    encloseSep
+      lbrace
+      rbrace
+      comma
+      (prettyDoc <$> M.keys deps)
