+++ src/Language/Mimsa/Types/ResolverError.hs
@@ -1,7 +1,17 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Language.Mimsa.Types.ResolverError where
 
+import qualified Data.Map as M
+import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Name
 
 data ResolverError = MissingBinding Name Bindings
   deriving (Eq, Ord, Show)
+
+instance Printer ResolverError where
+  prettyPrint (MissingBinding name (Bindings bindings')) = "A binding for " <> prettyPrint name <> " could not be found in " <> bindingsList
+    where
+      bindingsList = "[ " <> T.intercalate ", " (prettyPrint <$> M.keys bindings') <> " ]"
