+++ src/Language/Mimsa/Types/TypeConstructor.hs
@@ -1,5 +1,9 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Language.Mimsa.Types.TypeConstructor where
 
+import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.MonoType
 
@@ -9,3 +13,8 @@ data TypeConstructor
         tcTypeVars :: [MonoType],
         tcConsTypes :: [MonoType]
       }
+  deriving (Show)
+
+instance Printer TypeConstructor where
+  prettyPrint (TypeConstructor consName _tyTypeVars consTypes) =
+    prettyPrint consName <> " " <> T.intercalate " " (prettyPrint <$> consTypes)
