+++ src/Language/Mimsa/Types/UniVar.hs
@@ -4,8 +4,13 @@
 
 module Language.Mimsa.Types.UniVar where
 
+import Data.Text.Prettyprint.Doc
 import GHC.Generics
+import Language.Mimsa.Printer
 
 newtype UniVar = UniVar Int
   deriving stock (Eq, Ord, Generic)
   deriving newtype (Show, Num)
+
+instance Printer UniVar where
+  prettyDoc (UniVar a) = pretty a
