+++ src/Language/Mimsa/Types/Usage.hs
@@ -1,5 +1,9 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Language.Mimsa.Types.Usage where
 
+import qualified Data.Text as T
+import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Name
 
@@ -8,6 +12,14 @@ data Usage
   | Direct Name ExprHash
   deriving (Eq, Ord, Show)
 
+instance Printer Usage where
+  prettyPrint (Transient name _) =
+    "Transient dependency of "
+      <> prettyPrint name
+  prettyPrint (Direct name _) =
+    "Direct dependency of "
+      <> prettyPrint name
+
 ----------
 
 data UsageError
@@ -15,3 +27,11 @@ data UsageError
   | CouldNotFindBinding Name
   | CouldNotFindStoreExpression ExprHash
   deriving (Eq, Ord, Show)
+
+instance Printer UsageError where
+  prettyPrint (CouldNotResolveDeps names) =
+    "Could not resolve deps: " <> T.intercalate ", " (prettyPrint <$> names)
+  prettyPrint (CouldNotFindBinding name) =
+    "Could not find binding " <> prettyPrint name
+  prettyPrint (CouldNotFindStoreExpression exprHash) =
+    "Could not find store expression for hash " <> prettyPrint exprHash
